src/websocket_tests/requirements.txt
eventlet
python-socketio
python-socketio[client]
python-dotenv
psycopg2